Refactor big components + Optimize performance for networks
