+++ website/src/components/index.ts
@@ -5,6 +5,9 @@ export * from './DocSearch'
 export * from './Heading'
 export * from './Image'
 export * from './Navigation'
+export * from './NetworkDetailsPage'
+export * from './NetworkPage'
+export * from './supported-networks'
 export * from './Table'
 export * from './TimeIcon'
 export * from './VideoEmbed'
