+++ website/src/pages/[locale]/supported-networks/[id].mdx
@@ -17,7 +17,11 @@ export const getStaticPaths = async () => {
     }
   }
 
-return { paths, fallback: false, } }
+  return { 
+    paths, 
+    fallback: false 
+  }
+}
 
 export const getStaticProps = async ({ params }) => {
   const { locale, id } = params
@@ -26,6 +30,7 @@ export const getStaticProps = async ({ params }) => {
       notFound: true,
     }
   }
+  
   const networks = await getSupportedNetworks()
   const network = networks.find((n) => n.id === id)
   if (!network) {
@@ -33,6 +38,7 @@ export const getStaticProps = async ({ params }) => {
       notFound: true,
     }
   }
+  
   const rawMdx = `---
 title: ${network.id}
 hideTableOfContents: true
@@ -45,6 +51,7 @@ parentRouteKey: supported-networks
 `
   const mdx = await buildDynamicMDX(rawMdx, { codeHighlight: false })
   const { pageMap } = await import(`.next/static/chunks/nextra-page-map-${locale}.mjs`)
+  
   return {
     props: {
       ...mdx,
@@ -56,10 +63,7 @@ parentRouteKey: supported-networks
     },
   }
 }
+
 export default function NetworkDetailPage({ network }) {
-  return (
-    <>
-      <NetworkPage network={network} />
-    </>
-  )
+  return <NetworkPage network={network} />
 }
